+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Update createPulseRouter to include dorkHome parameter
 ---
 
 ## [0.17.2] - 2026-03-19